xmhf: xmhf-hic: pass BSP indicator along with cpuid to xmhfhic_arch_relinquish _control_to_init_slab

+++ xmhf-hic/arch/x86-vmx-x86pc/xmhf-hic-x86vmxx86pc-code.c
@@ -2612,7 +2612,7 @@ void xmhfhic_arch_relinquish_control_to_init_slab(u64 cpuid){
     }
 
 
-    _XDPRINTF_("%s[%u]: Now in ring-3. Halting!\n", __FUNCTION__, cpuid);
+    _XDPRINTF_("%s[%u]: Now in ring-3. Halting!\n", __FUNCTION__, (u32)cpuid);
     HALT();
